src/main/java/org/example/Number.java
package org.example;

public class Number {

    public static boolean isPrime(int number) {
        if (number <= 1) {
            return false;
        }
        for (int i = 2; i <= Math.sqrt(number); i++) {
            if (number % i == 0) {
                return false;
            }
        }
        return true;
    }

    public static boolean isEven(int number) {
        return number % 2 == 0;
    }

    public static boolean isMultiple(int number, int divisor) {
        return number % divisor == 0;
    }

}
